# cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Микросервис подписок",
        "title": "Подписки",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/subs/total": {
            "get": {
                "description": "Подсчет стоимости подписок за период и по параметрам",
                "produces": [
                    "application/json"
                ],
                "summary": "Подсчет стоимости подписок",
                "parameters": [
                    {
                        "type": "string",
                        "description": "serviceName",
                        "name": "serviceName",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "startDate",
                        "name": "startDate",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "endDate",
                        "name": "endDate",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_helf4ch_gocrudl_internal_application.AppResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "$ref": "#/definitions/github_com_helf4ch_gocrudl_internal_dto.TotalSubscriptionResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_helf4ch_gocrudl_internal_application.AppResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_helf4ch_gocrudl_internal_application.AppResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/subs/{id}": {
            "get": {
                "description": "Получить подписку по id",
                "produces": [
                    "application/json"
                ],
                "summary": "Получить подписку",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_helf4ch_gocrudl_internal_application.AppResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "$ref": "#/definitions/github_com_helf4ch_gocrudl_internal_dto.ReadSubscriptionResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_helf4ch_gocrudl_internal_application.AppResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_helf4ch_gocrudl_internal_application.AppResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_helf4ch_gocrudl_internal_application.AppResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "description": "Редактировать подписку по id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Редактировать подписку",
                "parameters": [
                    {
                        "description": "новое тело подписки",
                        "name": "sub",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_helf4ch_gocrudl_internal_dto.UpdateSubscriptionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_helf4ch_gocrudl_internal_application.AppResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "$ref": "#/definitions/github_com_helf4ch_gocrudl_internal_dto.UpdateSubscriptionResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_helf4ch_gocrudl_internal_application.AppResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_helf4ch_gocrudl_internal_application.AppResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_helf4ch_gocrudl_internal_application.AppResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "description": "Удалить подписку по id",
                "produces": [
                    "application/json"
                ],
                "summary": "Удалить подписку",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_helf4ch_gocrudl_internal_application.AppResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "$ref": "#/definitions/github_com_helf4ch_gocrudl_internal_dto.DeleteSubscriptionResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_helf4ch_gocrudl_internal_application.AppResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_helf4ch_gocrudl_internal_application.AppResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_helf4ch_gocrudl_internal_application.AppResponse": {
            "type": "object",
            "properties": {
                "body": {},
                "error": {
                    "type": "string"
                },
                "ok": {
                    "type": "boolean"
                }
            }
        },
...
        "github_com_helf4ch_gocrudl_internal_dto.DeleteSubscriptionResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_helf4ch_gocrudl_internal_dto.ReadSubscriptionResponse": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "endDate": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "serviceName": {
                    "type": "string"
                },
                "startDate": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
...
        "github_com_helf4ch_gocrudl_internal_dto.TotalSubscriptionResponse": {
            "type": "object",
            "properties": {
                "cost": {
                    "type": "integer"
                }
            }
        },
        "github_com_helf4ch_gocrudl_internal_dto.UpdateSubscriptionRequest": {
            "type": "object",
            "properties": {
                "endDate": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "serviceName": {
                    "type": "string"
                },
                "startDate": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "github_com_helf4ch_gocrudl_internal_dto.UpdateSubscriptionResponse": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "endDate": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "serviceName": {
                    "type": "string"
                },
                "startDate": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        }
    }
}